Refatorei os métodos separando eles em outra classe
