+++ src/views/internal/student/class/add/components/ClassFormCodeReader.vue
@@ -20,7 +20,7 @@ export default {
             }
 
             this.value.code = decodedString
-            this.$emit('add')
+            this.$emit('addClass')
         },
         isCodeValid(decodedString) {
             return decodedString.length > 0;
